requirements.txt
aiogram>=3.0.0
tortoise-orm>=0.20.0
python-dotenv>=1.0.0
openai>=1.0.0
asyncpg>=0.29.0
